now user can place order multiple items with quantities,

+++ src/main/java/com/example/MyMart/DTO/Request/ProductRequest.java
@@ -10,6 +10,7 @@ import lombok.*;
 @Builder
 public class ProductRequest {
     private  String name;
-    private double price;
+    private int price;
     private Category category;
+
 }
